Add a topic field via incoming processor
If the directory name containing the assets contains a circumflex (^)
then what precedes it is assigned to the "topic" field of the CouchDB
document. Any underscores within the topic value are replaced with
spaces. The rest of the directory name is treated as before: set of tags
separated by underscores, followed by optional at sign (@) and
location (whose value may contain underscores, which are replaced by
spaces).

make ct passes