No hard-coded checksums in the test suite
Because the images are now auto-oriented, the image data may vary
slightly from platform to platform, and hence the checksum may vary.
Instead of using hard-coded checksums, retrieve the checksum based on
the file name. While not at all efficient, it is okay for the purpose of
testing.

rebar3 ct passes