Rip out all traces of the web interface
The primitive web interface needs to be replaced with something better.
Using Nitrogen is hard, especially compared to a framework like Phoenix
which has a very modern design. Nitrogen is a lot like building your own
computer from parts, without a manual, where some of the parts are made
by you. Phoenix is extremely easy to get started with and very easy to
add new routes, pages, controllers, etc.

rebar3 ct passes

Below are notes on the available options in order of preference. It
should be obvious from this that Phoenix will become the front-end for
tanuki going forward.

a) Elixir for the web, Erlang for the rest

* [Phoenix](http://www.phoenixframework.org/) for the user interface
    - Easy to understand, featureful, familiar design, well maintained
    - Architecture is amenable to backend developers (no JavaScript)
* Already have the incoming processor and backend written in Erlang
* Call into the tanuki_backend using gen_server calls
    - Caching in the backend is easy
    - Maintains separation of concerns

b) All Elixir

* Phoenix and CouchDB can be made to work together
    - https://github.com/locolupo/couchdb_connector
    - https://github.com/mkrogemann/bird_watch (demo app)
* However, we already have a bunch of working code in Erlang...

c) All Erlang

* Nitrogen is falling behind, latest release was in February 2015
* ChicagoBOSS is falling behind, last commit was in June 2016
* Cowboy is well maintained, but only suitable for REST-based services
* There really isn't much else that is actively maintained