Switch from rebar to mix
It seems that getting rebar to build a release that contains a mix of
Erlang and Elixir applications is not feasible at this time. As such,
switch to using the mix tool (provided with Elixir), which does handle
this particular situation.

This is preparing for the introduction of the web interface, which will
be written in Elixir, based on Phoenix.

Notes on the selection process:

* How to get `rebar3` to build mix-based apps?
    * tried pre/post hooks, didn't do anything
    * tried setting `lib_dirs`, didn't do anything
    * looked at rebar plugins for compiling `.ex` files
        * https://github.com/yrashk/rebar_elixir_plugin
        * https://github.com/barrel-db/rebar3_elixir_compile
        * they only know how to compile Elixir, not run `mix`
* Maybe use `erlang.mk` instead of rebar?
    * Has plugins for hex and mix (and rust and other languages)
    * It's all just glorified Makefiles, so yuck
    * Not an official anything (both rebar and mix are official build tools)