Cache results from by_tags/1, add by_tags/2
Results from the request for assets with a given set of tags are cached
by those set of tags. Repeated requests will return the same results.
Only one set of results is retained, so a request for a different set of
tags will clear the cache.

Add a new function by_tag/2 that takes a sorting function so that the
results from the query can be sorted before they are cached.

mix ct passes