Remove all use of emagick.rs module
Compiling magick-rust and emagick.rs is becoming increasingly difficult as
the rust-bindgen tool continues to evolve. At this point, it is impossible
to produce bindings for MagickWand (i.e. ImageMagick) using the latest
bindgen tools. The simpler strategy is to instead use the erlang-exif module
and the ImageMagick `convert` binary.

Since `convert` likes to read and write files on disk, the generated
thumbnails and previews are now stored on disk. Every 5 minutes the
server will prune the generated thumbnails, removing those least recently
accessed. If the fstat of many files proves to be slow, some caching can be
introduced. For now it seems to be working fine.

mix test passes