Handle range header properly for video playback
Need to handle the Range request header so that Safari will play videos
properly in the browser. Use the video tag on the asset detail page to have
the video play back directly from there. This may also help with video
playback in Google Chrome. Also need to modify the mimetype for the video
tag, converting "video/quicktime" to "video/mp4" so the video will play in
Chrome. Note that the content-type header is still the original value.

mix test passes