Use bootstrap for thumbnails and pagination
Use Bootstrap for better styling of the thumbnails and their captions. Same
for the pagination links, which look much better than before.

Add "..." where there is a gap between first numbered page and page 1, and
last numbered page and last page, so it is obvious that there are additional
pages inbetween.

mix test passes