Use emagick_rs once again for image handling
Generate the thumbnails in memory using emagick_rs, storing in mnesia for
the purpose of caching. Preview images are always generated, not cached.
Nothing is cached on disk any more. All of this is just as before. One
change is that the mime type is discovered rather than hard-coded to
"image/jpeg" in the backend.

mix test passes