Attempt to support HEIC files
However, ImageMagick, and hence magick-rust and emagick.rs, does not yet
know about this file format. For now, the changes have been made and tests
are ready to uncomment when the time comes. There is a patch for ImageMagick
but it will be some time before it gets released (see issue #507 on GitHub).

mix test passes