Invoke ffmpeg with explicit output options
The fluent-ffmpeg API offers high-level functions, such as screenshots(),
but unfortunately it takes away much of the control needed to get desirable
results. In particular, we want to maintain the aspect ratio when creating a
thumbnail for a video. And JPG format is much smaller than PNG, but the
screenshots() function insists on writing PNG.

npm test passes