Add thumbnail caching
The test is performed by querying the log file, which we have in place
already, and it is configured correctly, so querying is easy.

Also, "path" does not make for a good local variable name as it shadows the
name typically used for importing the 'path' module.

npm test passes