Allow updating tag letter case
Keep only those old tags that do not match the incoming tags in a case
insensitive manner. That allows for the update to change the letter case of
a tag. Since our index is case insensitive, it only makes sense to keep one
version of the tag.

npm test passes