Prefix checksum value with hash function name
The checksum value in the document must now be prefixed with a hash function
name, and hyphen separator. For instance, "sha256-cafebabe". The GraphQL
resolver ensures there is a hyphen in the checksum argument to the lookup
field. The migration code also has renamed the "sha256" field to "checksum"
since the algorithm is no longer part of the schema.

npm test passes