Update some module dependencies, improve logging
Upgrade to PouchDB 7.0 which had no significant API changes. Also change the
logging to not use the default logger instance; also upgrade winston to the
latest release. This requires a hack to work around a bug in the query
feature used in one of the tests.

Also use NODE_PATH instead of app-module-path to load our own modules.

npm test passes