Translate TypeScript back to JavaScript
There is almost no benefit to using TypeScript, and it introduces more work
and complication to the setup and running of the application. Eventually all
of the JavaScript code is going to be replaced with Rust anyway.

Move all of the other JavaScript code to the src directory, so all of the
JavaScript code is in one directory.

Move all of the ReasonML code to the src directory as well, since it is also
source code, and the compilers and interpreters are not at all confused by
the presence of files that do not match the expected pattern.

npm test passes