Add proper GraphQL server support
The current schema has been converted to use Dart types and the GraphQL
support in the Angle mono-repo. Using custom types, however, the GraphiQL
interface is a little broken, and cannot display the schema. However, the
queries still function normally. Most of the functionality has not been
tested yet, and the object field resolvers may or may not really do
anything, but it is a good start.

flutter test passes