Add usecase for updating an asset
The logic in the GraphQL resolver implementation written in JavaScript has
been incorporated into this usecase. This usecase takes the asset and an
"asset input" record and essentially does what the resolver did.

However, the case-insensitive merging of caption tags over existing tags was
removed, as it seemed unnecessarily complicated. Instead, the input tags
replace the asset tags, if given, and then the caption tags are added to
those tags. The result is always sorted and de-duplicated.

cargo test passes