Perform lowercase of keys consistently in queries
The search usecase will now lowercase the inputs and index values when
performing filtering, and only for filtering. Similarly, the data source
performs lowercasing of the index keys _and_ the query parameters when
performing a query. This keeps the behavior, or algorithm, in each module
consistent within that module, without any dependency (or connascence of
algorithm, the worst kind) between modules.

cargo test passes