Add HTTP server and file upload endpoint
Introducing actix-web and actix-multipart to serve the web interface and
allow for file uploads. For now the files are placed in ./tmp in the current
directory, but eventually they will be placed in the repository. Since the
multipart crate performs the media type detection, the mime_guess crate has
been removed for now. The import usecase now expects the media type to be
specified when adding the new asset.

cargo test passes