Add count query for GraphQL schema
Use an Arc to hold the data source in the record repository, which makes it
easy to keep a reference to the data source (as a dynamic trait) in the
GraphQL context, and also create a copy to pass to the new record
repository (which in turn gets passed to the usecase). This avoids weird
cloning tricks, and keeps the mocking simple in the tests.

cargo test passes