Read image dimensions and save to asset record
Add a new dimensions field to the asset struct, which holds in the pixel
width and height of the (image) asset. Is None if the asset is not an image.
Note that images which are lying sideways will have flipped dimensions.

Added non-consuming builder functions to make it easier to create an Asset
in the future, without having to name every field.

cargo test passes