Append the correct extension to the asset id
If the caller provides a filename whose extension does not match what would
be expected for the media type given, then append one of the corresponding
extensions (defaults to the first one in the list). To be precise, the media
type is guessed from the given filename extension, and if that type differs
from the provided media type, then the first extension associated with the
provided media type is appended to the end of the asset identifier.

This will allow for the blob repository to correctly guess the media type
using only the asset identifier, and avoid querying the database.

This is only really helpful if the client sending the file can correctly
guess the media type despite the filename having a strange extension.

cargo test passes