Use default jpeg quality factor of 0.75
This is the value that libvips uses by default. It reduces the size of the
thumbnail considerably by reducing the quality just a bit, without venturing
too far into the low quality end of the spectrum.

Regarding libvips, tried that and it has some issues. First, the wrapper
forces passing the import and export ICC profile file names to the thumbnail
with options function, which makes it inconvenient. More importantly, the
library crashes when creating thumbnails for some images.

Clean up the todo wrt image orientation, which has already been settled.

cargo test passes