Remove the duration field from the asset record
It is not all that useful really and it just clutters the struct with a
field that does not apply to most assets, the majority of which are images.

This also helps to reduce the need for a video parsing library (e.g. ffmpeg
or ffprobe), which seems to be a challenge in Rust.

cargo test passes