Use video tag for asset thumbnails everywhere
Saves the trouble of generating thumbnails for videos which is more than
difficult with the available libraries in Rust. Yes, could just invoke
ffmpeg on the video, but there is no need to bother when the video tag works
perfectly. Except for Safari, which interprets the click to play as click to
open the asset preview page.