Add ingest usecase to import all uploads files
Usecase that imports anything in the "uploads" folder, detecting the media
type from the file name, creating a new record, and moving the file into the
blob repository. Added button on the upload screen to invoke the new GraphQL
mutation to perform the work.

Also updated the GraphQL schema for the frontend and fixed all of the errors.

cargo test passes