Change selector dropdown to toggle items
This is a work-around for the issue of the dropdown covering the chips
input, sometimes making it impossible to click the (x) to remove a chip.
That will eventually be resolved in a different way, but for now the
dropdown has been changed to toggle the selection rather than always
appending to the list.

flutter test passes