fix: split date range selector into two fields
Change the date range selector on the asset browser screen to be two
separate fields for "after" and "before" dates. This works much better than
the date range picker that takes over the entire screen and is very slow to
scroll (it's trying to show a wide range of dates in calendar form).

The new fields are an improvement, however, there is no button to reset the
form fields, so once dates are input, they cannot be cleared. Should add an
overlay "close" icon on each field to permit clearing them individually.

fvm flutter test passes