fix: repair missing filename and media type
If the asset file name has been cleared (for example, due to a bug in the
application), then diagnose and repair will be hampered. This is worsened if
the media type field was also cleared (due to a bug), making the assets
irreparable. Wit this change, the repair will now take the asset ID, which
is the original path encoded as base64, and use that (generated) file name
as the asset file name, thereby enabling the media type repair to succeed.
The original file name is still lost forever, but better to have something
than nothing at all.

cargo test passes