Change date selector to year + season dropdown
Model the date range selector after the AniList.co anime search form,
which presents a dropdown of years and another dropdown of seasons. This is
much easier to operate than the date range selector that invokes the system
date selector calendar widget. Clearing the selections is also one click
instead of two (one to clear and the other to dismiss the calendar dialog
that pops up because the value was cleared).

npm test passes