feat: streamline asset import process
Improve the recent imports page significantly such that it uses an input
chips field to specify tags and a typeahead field to specify the location.
The assets are shown in a grid view and allow multiple selections. Recents
page now has pagination to avoid loading all of the recent assets (was only
showing 100 but the query still returned everything).

cargo/flutter test passes