fix: allow new location despite others with prefix
If the entered location on the import page happens to be the prefix of an
existing location in the database, you cannot select the query as entered.
Now that is fixed, allowing for a new location ("hakone") when an existing
location with a matching prefix ("hakona gardens") already exists.

fvm flutter test passes