feat: CSV in locations will be indexed separately
On the imports screen, if a location is entered as "tokyo, japan" then the
index will contain entries for "tokyo" and "japan" in the all_locations
view, and "tokyo, japan" in the raw_locations view. This enables the import
screen to allow completion on "tokyo, japan" while the search screen will
offer "tokyo" and "japan" as separate locations.

cargo test passes