feat: use reverse geocoding to locate new assets
Use a reverse geocoding service, if certain environment variables are set,
to look up the city and region where an asset was created. This works by
reading the GPS location from the Exif data in images and passing those
values to a mapping API to produce a location. Initially only supports
Google Maps API, retrieving the "locality" and "administrative_area_level_1"
values for a location.

Also, the location property of an asset may now have 3 possible values:
null, a string ("label"), or a struct consisting of "label", "city", and
"region". When serializing the record, if city and region are both unset,
then only the string label is written. If all 3 fields are unset, then null
is written.

cargo test passes