feat: add dump/load features for crude backup
New GraphQL operations to dump the database records to a file, with each
asset record formatted in JSON and written as one line. The load operation
will read the given file, line-by-line, deserializing the asset records and
overwriting whatever is in the database. This allows for a simple backup and
restore as well as large-scale bulk edits by modifying the JSON file before
loading back into the database.

cargo test passes