feat: parse text into fully-qualified location
On the imports screen, if the location text entered by the user matches a
certain format, parse it into a fully-qualified location record. That is, if
the value is of the form "city, region" then a location with city and region
is created. If the text is "label - city, region" then a location will all
fields populated is created.

flutter test passes