feat: user can now upload replacement asset
En lieu of image/video editing features within the application, the user can
download the raw asset and upload a replacement from both the asset details
screen and the asset edit screen. This can also be used to provide an asset
if one went missing. In either case, a new record is created that merges the
old data with the new filename, length, media type, as well as assigning a
new identifier for the asset. Then the old asset and record are removed.

cargo/flutter test passes