fix: missing city, promote other location fields
If the geocoded location is missing the city but has region and country,
then promote those two fields to be the city and region. The domain entity
does not have a country field and ignoring the geocoded data completely
because the city is missing is not helpful.

cargo test passes