fix: update use case handling of custom date
The custom date can no longer be cleared; it is unreasonable to expect the
client to always set the custom date on every request, even if the user is
only changing some other attribute. As such, once a custom date is set, it
can be modified but it cannot be removed.

cargo leptos test passes