fix: do not wipe out original date on replace
When replacing an asset, the original date/time will only be updated if the
new asset has its own original date/time value. For instance, the old video
format had an original date/time, but the converted video did not, and as a
result the record would lose the original date/time.

cargo leptos test