feat: add lightbox style perusal of search results
From the browse or search pages, clicking on an asset will show the asset
details page. On the left and right side of the asset display there are now
arrow buttons. These will move backward and forward through the set of
results from the browse or search page.

cargo leptos test passes