fix: remove old comma-based location parsing
Before the improvements to the location information in the entity records,
the use of commas was a way of allowing for multiple values that could be
matched independently. For example, "Paris, France" would result in two rows
in the index. However, with the improved location data, a string like that
would be parsed as the city and region rather than a long label only value.

cargo test passes