feat: support location search on specific fields
The advanced query feature now supports "loc:city:..." to match on the city
field of the location, as well as "label" and "region" to match on the
corresponding fields. Using "loc:any" is the same as "loc:" without the
label/city/region specifier. That is, if the `loc` predicate has two colons,
then the first value must be one of `any`, `label`, `city`, or `region` and
the second value is the query string; if `loc` is followed by a single
colon, the predicate will assume `any` field when matching the query.

Examples:

- loc:beach -> as before, any field equals 'beach'
- loc:label:beach -> matches only when label is 'beach'
- loc:city:paris -> matches only when city is 'paris'
- loc:region:france -> matches only when region is 'france'
- loc:any:beach -> matches if any field equals 'beach'

cargo test passes