fix: use only strong reference to database object
Since the beginning, the database references have been retained using a Weak
reference, "to allow the database to eventually be closed". Since there have
been seemingly random errors when trying to open the database, this change
should reduce, if not entirely eliminate them.

Based on testing on production with additional logging, the database
references are being recreated quite frequently. This suggests that the
database is being opened and closed many times. After this change was
deployed, the difference in speed was very significant. The conclusion is
that the use of the Weak reference was more harmful than helpful and served
no real purpose.

cargo leptos test passes