fix: remove unused query-by-filename feature
Searching for assets by filename was not exposed in the interface and it
will not be missed. The value of searching for assets by filename is very
doubtful as it would require remembering the one attribute of an asset that
is the least interesting and relevant. Most of the time the filenames are
generated (IMG_1234.JPG) or cryptic (20150419171116-63EK7JXW-P1510081.jpg).

cargo leptos test passes