fix: retain current location if new value is none
The behavior and tests were wrong in that an existing location value was
wiped out if the input location was None. This can happen when replacing an
asset with a new file and that file does not have any GPS data. As a result,
any location values in the asset record were lost.

cargo leptos test passes