fix: restore custom Location serialization
Removing the custom serialization logic for assets and locations was
premature, as the existing JSON format produced by the dump operation
requires the deserializer used by the load operation to understand the
concise representation of a Location that has only a label field.

Changed the RocksDB asset serde code to perform the same optimization when
storing a location that contains only a label; it will save space since
there are many assets that have only a location label.

cargo leptos test passes